feat: Check type of the params of #initialize
